+++ src/main/integrations/child-process.ts
@@ -18,11 +18,18 @@ export interface ChildProcessOptions extends NodeChildProcessOptions {
   breadcrumbs: Readonly<ExitReason[]>;
   /** Child process events that generate Sentry events */
   events: Readonly<ExitReason[]>;
+  /**
+   * Whether to also capture Sentry logs for child process events
+   *
+   * default: false
+   */
+  logs: boolean;
 }
 
 const DEFAULT_OPTIONS: ChildProcessOptions = {
   breadcrumbs: EXIT_REASONS,
   events: ['abnormal-exit', 'launch-failed', 'integrity-failure'],
+  logs: false,
 };
 
 type LogFn = (msg: ParameterizedString, attributes: Log['attributes']) => void;
@@ -63,6 +70,7 @@ export const childProcessIntegration = defineIntegration((userOptions: Partial<O
   const options: ChildProcessOptions = {
     breadcrumbs: Array.isArray(breadcrumbs) ? breadcrumbs : breadcrumbs === false ? [] : DEFAULT_OPTIONS.breadcrumbs,
     events: Array.isArray(events) ? events : events === false ? [] : DEFAULT_OPTIONS.events,
+    logs: !!userOptions.logs,
   };
 
   return {
@@ -76,7 +84,7 @@ export const childProcessIntegration = defineIntegration((userOptions: Partial<O
       // only hook these events if we're after more than just the unresponsive event
       if (allReasons.length > 0) {
         const clientOptions = client.getOptions() as ElectronMainOptions;
-        const enableLogs = !!clientOptions?.enableLogs;
+        const logs = options.logs;
 
         app.on('child-process-gone', (_, details) => {
           const { reason } = details;
@@ -97,7 +105,7 @@ export const childProcessIntegration = defineIntegration((userOptions: Partial<O
               data: details,
             });
 
-            if (enableLogs) {
+            if (logs) {
               log(messageFmt, {
                 'sentry.origin': 'auto.electron.child-process',
                 exitCode: details.exitCode,
@@ -128,7 +136,7 @@ export const childProcessIntegration = defineIntegration((userOptions: Partial<O
               data: details,
             });
 
-            if (enableLogs) {
+            if (logs) {
               log(messageFmt, {
                 'sentry.origin': 'auto.electron.child-process',
                 exitCode: details.exitCode,